+++ src/main/java/com/lmsltirollcallsjtu/common/bean/bo/SignRecordsBo.java
@@ -8,7 +8,7 @@ import lombok.experimental.Tolerate;
 @Builder
 public class SignRecordsBo {
     private String id;//表记录id
-    private Integer userCode;//学生学号
+    private Long userCode;//学生学号
     private String rollcallCode;//点名编号
     private String openId;//绑定微信编号
     private String state;//签到状态
